fix gcc 11 build errors

+++ src/dory/input_dg/input_dg_util.cc
@@ -46,8 +46,9 @@ TMsg::TPtr Dory::InputDg::BuildMsgFromDg(const void *dg, size_t dg_size,
     TMsgStateTracker &msg_state_tracker) {
   assert(dg);
   const auto *dg_bytes = reinterpret_cast<const uint8_t *>(dg);
-  size_t fixed_part_size = INPUT_DG_SZ_FIELD_SIZE +
-      INPUT_DG_API_KEY_FIELD_SIZE + INPUT_DG_API_VERSION_FIELD_SIZE;
+  size_t fixed_part_size = size_t(INPUT_DG_SZ_FIELD_SIZE) +
+      size_t(INPUT_DG_API_KEY_FIELD_SIZE) +
+      size_t(INPUT_DG_API_VERSION_FIELD_SIZE);
 
   if (dg_size < fixed_part_size) {
     DiscardMalformedMsg(dg_bytes, dg_size, anomaly_tracker, log_discard);
@@ -62,7 +63,8 @@ TMsg::TPtr Dory::InputDg::BuildMsgFromDg(const void *dg, size_t dg_size,
   }
 
   int16_t api_key = ReadInt16FromHeader(dg_bytes + INPUT_DG_SZ_FIELD_SIZE);
-  size_t key_part_size = INPUT_DG_SZ_FIELD_SIZE + INPUT_DG_API_KEY_FIELD_SIZE;
+  size_t key_part_size = size_t(INPUT_DG_SZ_FIELD_SIZE) +
+      size_t(INPUT_DG_API_KEY_FIELD_SIZE);
   int16_t api_version = ReadInt16FromHeader(dg_bytes + key_part_size);
   const uint8_t *versioned_part_begin = &dg_bytes[fixed_part_size];
   const uint8_t *versioned_part_end = versioned_part_begin +
